+++ iOS/Plugins/FlipperKitLayoutPlugin/FlipperKitLayoutComponentKitSupport/SKComponentLayoutDescriptor.mm
@@ -24,7 +24,7 @@
 #import <ComponentKit/CKInsetComponent.h>
 
 #import <FlipperKitHighlightOverlay/SKHighlightOverlay.h>
-#import <FlipperKitLayoutPlugin/SKObject.h>
+#import <FlipperKitLayoutHelpers/SKObject.h>
 #import <FlipperKitLayoutTextSearchable/FKTextSearchable.h>
 
 #import "CKComponent+Sonar.h"