+++ iOS/Plugins/FlipperKitLayoutPlugin/FlipperKitLayoutPlugin/FlipperKitLayoutPlugin.h
@@ -7,14 +7,15 @@
 
 #if FB_SONARKIT_ENABLED
 
+#import "SKDescriptorMapper.h"
+
 #import <Foundation/Foundation.h>
 
 #import <FlipperKit/FlipperPlugin.h>
 #import <FlipperKit/SKMacros.h>
 
-#import "SKDescriptorMapper.h"
-#import "SKInvalidation.h"
-#import "SKTapListener.h"
+#import <FlipperKitLayoutHelpers/SKInvalidation.h>
+#import <FlipperKitLayoutHelpers/SKTapListener.h>
 
 @interface FlipperKitLayoutPlugin
     : NSObject<FlipperPlugin, SKInvalidationDelegate>