+++ react-native/ReactNativeFlipperExample/metro.config.js
@@ -12,7 +12,7 @@ module.exports = {
     getTransformOptions: async () => ({
       transform: {
         experimentalImportSupport: false,
-        inlineRequires: false,
+        inlineRequires: true,
       },
     }),
   },